Add k8s ttl controller flag for scaler (#41)

+++ k8s/agent-scaler/agent-scaler.sh
@@ -3,24 +3,30 @@
 batch_size=${BATCH_SIZE:-1}
 max_jobs=${MAX_JOBS:-10}
 check_delay=${CHECK_DELAY:-10}
+# Kubernetes 1.21 supports jobs TTL controller, which cleans up jobs automatically
+# see https://kubernetes.io/docs/concepts/workloads/controllers/ttlafterfinished/
+use_k8s_ttl_controller=${USE_K8s_TTL_CONTROLLER:-"false"}
 
 namespace=$(cat /var/run/secrets/kubernetes.io/serviceaccount/namespace)
 
 # run forever
 while true; do
-    # cleanup finished jobs (status 1/1)
-    for job in $(kubectl get jobs -n "${namespace}" | grep -E -e 'joshua-agent-[0-9-]*\s*1/1' | awk '{print $1}'); do
-        echo "=== Job $job Completed ==="
-        kubectl delete job "$job" -n "${namespace}"
-    done
 
-    # cleanup failed jobs
-    # pod name is always prefixed with the job name
-    # e.g. "joshua-agent-XXXXXXXXXXXX-XX-yyyyy"
-    for job in $(kubectl get pods -n "${namespace}" | grep -E -e "Completed" -e "Error" | cut -f 1-4 -d '-') ; do
-        echo "=== Deleting Job $job ==="
-        kubectl delete job "$job" -n "${namespace}"
-    done
+    if [ "${use_k8s_ttl_controller}" == "false" ] ; then
+      # cleanup finished jobs (status 1/1)
+      for job in $(kubectl get jobs -n "${namespace}" | grep -E -e 'joshua-agent-[0-9-]*\s*1/1' | awk '{print $1}'); do
+          echo "=== Job $job Completed ==="
+          kubectl delete job "$job" -n "${namespace}"
+      done
+
+      # cleanup failed jobs
+      # pod name is always prefixed with the job name
+      # e.g. "joshua-agent-XXXXXXXXXXXX-XX-yyyyy"
+      for job in $(kubectl get pods -n "${namespace}" | grep -E -e "Completed" -e "Error" | cut -f 1-4 -d '-') ; do
+          echo "=== Deleting Job $job ==="
+          kubectl delete job "$job" -n "${namespace}"
+      done
+    fi
 
     # get the current ensembles
     num_ensembles=$(python3 /tools/ensemble_count.py)
